fix: move failures and autocallouts
