--- fbw-a32nx/src/systems/instruments/src/EFB/index.html
@@ -1,17 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-    <head>
-        <meta charset="utf-8">
-
-        <title>a32nx | EFB</title>
-
-        <meta name="viewport" content="width=device-width, initial-scale=1">
-
-        <script type="module" defer src="./index.tsx"></script>
-    </head>
-    <body>
-        <div id="ROOT_ELEMENT">
-            <div id="MSFS_REACT_MOUNT"></div>
-        </div>
-    </body>
-</html>
