+++ fbw-common/src/systems/instruments/src/EFB/failures-orchestrator-provider.tsx
@@ -2,9 +2,9 @@
 //
 // SPDX-License-Identifier: GPL-3.0
 
-import React, { useState } from 'react';
-import { A320Failure, Failure, FailuresOrchestrator } from '@failures';
-import { useUpdate } from '@flybywiresim/fbw-sdk';
+import React, { PropsWithChildren, useState } from 'react';
+import { Failure, FailuresOrchestrator, useUpdate } from '@flybywiresim/fbw-sdk';
+import { FailureDefinition } from '../../../shared/src/failures/failures-orchestrator';
 
 interface FailuresOrchestratorContext {
     allFailures: Readonly<Readonly<Failure>[]>,
@@ -14,105 +14,7 @@ interface FailuresOrchestratorContext {
     deactivate(identifier: number): Promise<void>;
 }
 
-const createOrchestrator = () => new FailuresOrchestrator('A32NX', [
-    [21, A320Failure.Acsc1Lane1, 'ACSC 1 Lane 1'],
-    [21, A320Failure.Acsc1Lane2, 'ACSC 1 Lane 2'],
-    [21, A320Failure.Acsc2Lane1, 'ACSC 2 Lane 1'],
-    [21, A320Failure.Acsc2Lane2, 'ACSC 2 Lane 2'],
-    [21, A320Failure.HotAir, 'Trim Air Pressure Regulating Valve'],
-    [21, A320Failure.TrimAirHighPressure, 'Trim Air System High Pressure'],
-    [21, A320Failure.CkptTrimAirFailure, 'Cockpit Trim Air Valve'],
-    [21, A320Failure.FwdTrimAirFailure, 'Forward Zone Trim Air Valve'],
-    [21, A320Failure.AftTrimAirFailure, 'Aft Zone Trim Air Valve'],
-    [21, A320Failure.CkptDuctOvht, 'Cockpit Duct Overheat'],
-    [21, A320Failure.FwdDuctOvht, 'Forward Zone Duct Overheat'],
-    [21, A320Failure.AftDuctOvht, 'Aft Zone Duct Overheat'],
-    [21, A320Failure.CabinFan1Failure, 'Cabin Fan 1'],
-    [21, A320Failure.CabinFan2Failure, 'Cabin Fan 2'],
-    [21, A320Failure.LabGalleyFan, 'Extraction Fan of lavatory and galley'],
-
-    [22, A320Failure.Fac1Failure, 'FAC 1'],
-    [22, A320Failure.Fac2Failure, 'FAC 2'],
-
-    [24, A320Failure.TransformerRectifier1, 'TR 1'],
-    [24, A320Failure.TransformerRectifier2, 'TR 2'],
-    [24, A320Failure.TransformerRectifierEssential, 'ESS TR'],
-    [24, A320Failure.StaticInverter, 'Static Inverter'],
-    [24, A320Failure.Generator1, 'Generator 1'],
-    [24, A320Failure.Generator2, 'Generator 2'],
-    [24, A320Failure.ApuGenerator1, 'APU Generator'],
-    [24, A320Failure.AlternatingCurrent1, 'AC 1'],
-    [24, A320Failure.AlternatingCurrent2, 'AC 2'],
-    [24, A320Failure.AlternatingCurrentEssential, 'AC ESS'],
-    [24, A320Failure.AlternatingCurrentEssentialShed, 'AC ESS SCHED'],
-    [24, A320Failure.AlternatingCurrentStaticInverter, 'AC STAT INV'],
-    [24, A320Failure.AlternatingCurrentGndFltService, 'AC GND FLT SRV'],
-    [24, A320Failure.DirectCurrent1, 'DC 1'],
-    [24, A320Failure.DirectCurrent2, 'DC 2'],
-    [24, A320Failure.DirectCurrentEssential, 'DC ESS'],
-    [24, A320Failure.DirectCurrentEssentialShed, 'DC ESS SCHED'],
-    [24, A320Failure.DirectCurrentBattery, 'DC BAT'],
-    [24, A320Failure.DirectCurrentHot1, 'DC HOT 1'],
-    [24, A320Failure.DirectCurrentHot2, 'DC HOT 2'],
-    [24, A320Failure.DirectCurrentGndFltService, 'DC GND FLT SRV'],
-
-    [27, A320Failure.Elac1Failure, 'ELAC 1'],
-    [27, A320Failure.Elac2Failure, 'ELAC 2'],
-    [27, A320Failure.Sec1Failure, 'SEC 1'],
-    [27, A320Failure.Sec2Failure, 'SEC 2'],
-    [27, A320Failure.Sec3Failure, 'SEC 3'],
-    [27, A320Failure.Fcdc1Failure, 'FCDC 1'],
-    [27, A320Failure.Fcdc2Failure, 'FCDC 2'],
-
-    [29, A320Failure.GreenReservoirLeak, 'Green reservoir leak'],
-    [29, A320Failure.BlueReservoirLeak, 'Blue reservoir leak'],
-    [29, A320Failure.YellowReservoirLeak, 'Yellow reservoir leak'],
-    [29, A320Failure.GreenReservoirAirLeak, 'Green reservoir air leak'],
-    [29, A320Failure.BlueReservoirAirLeak, 'Blue reservoir air leak'],
-    [29, A320Failure.YellowReservoirAirLeak, 'Yellow reservoir air leak'],
-    [29, A320Failure.GreenReservoirReturnLeak, 'Green reservoir return leak'],
-    [29, A320Failure.BlueReservoirReturnLeak, 'Blue reservoir return leak'],
-    [29, A320Failure.YellowReservoirReturnLeak, 'Yellow reservoir return leak'],
-    [29, A320Failure.GreenEdpOverheat, 'Green engine pump overheat'],
-    [29, A320Failure.BlueEpumpOverheat, 'Blue electric pump overheat'],
-    [29, A320Failure.YellowEdpOverheat, 'Yellow engine pump overheat'],
-    [29, A320Failure.YellowEpumpOverheat, 'Yellow electric pump overheat'],
-
-    [31, A320Failure.LeftPfdDisplay, 'Captain PFD display'],
-    [31, A320Failure.RightPfdDisplay, 'F/O PFD display'],
-
-    [32, A320Failure.LgciuPowerSupply1, 'LGCIU 1 Power supply'],
-    [32, A320Failure.LgciuPowerSupply2, 'LGCIU 2 Power supply'],
-    [32, A320Failure.LgciuInternalError1, 'LGCIU 1 Internal error'],
-    [32, A320Failure.LgciuInternalError2, 'LGCIU 2 Internal error'],
-
-    [32, A320Failure.GearProxSensorDamageGearUplockNose1, 'Proximity sensor damage uplock nose gear #1'],
-    [32, A320Failure.GearProxSensorDamageGearDownlockNose2, 'Proximity sensor damage downlock nose gear #2'],
-    [32, A320Failure.GearProxSensorDamageGearUplockRight1, 'Proximity sensor damage uplock right gear #1'],
-    [32, A320Failure.GearProxSensorDamageGearDownlockRight2, 'Proximity sensor damage downlock right gear #2'],
-    [32, A320Failure.GearProxSensorDamageGearUplockLeft2, 'Proximity sensor damage uplock left gear #2'],
-    [32, A320Failure.GearProxSensorDamageGearDownlockLeft1, 'Proximity sensor damage downlock left gear #1'],
-    [32, A320Failure.GearProxSensorDamageGearDoorClosedNose1, 'Proximity sensor damage closed nose gear door #1'],
-    [32, A320Failure.GearProxSensorDamageGearDoorOpenedNose2, 'Proximity sensor damage opened nose gear door #2'],
-    [32, A320Failure.GearProxSensorDamageGearDoorClosedRight2, 'Proximity sensor damage closed right gear door #2'],
-    [32, A320Failure.GearProxSensorDamageGearDoorOpenedRight1, 'Proximity sensor damage opened right gear door #1'],
-    [32, A320Failure.GearProxSensorDamageGearDoorClosedLeft2, 'Proximity sensor damage closed left gear door #2'],
-    [32, A320Failure.GearProxSensorDamageGearDoorOpenedLeft1, 'Proximity sensor damage opened left gear door #1'],
-
-    [32, A320Failure.GearActuatorJammedGearNose, 'Nose gear jammed actuator'],
-    [32, A320Failure.GearActuatorJammedGearLeft, 'Main left gear jammed actuator'],
-    [32, A320Failure.GearActuatorJammedGearRight, 'Main right gear jammed actuator'],
-    [32, A320Failure.GearActuatorJammedGearDoorNose, 'Nose gear door jammed actuator'],
-    [32, A320Failure.GearActuatorJammedGearDoorLeft, 'Main left gear door jammed actuator'],
-    [32, A320Failure.GearActuatorJammedGearDoorRight, 'Main right gear door jammed actuator'],
-
-    [32, A320Failure.GreenBrakeHydraulicLeak, 'Green brakes circuit leak'],
-    [32, A320Failure.YellowBrakeHydraulicLeak, 'Yellow brakes circuit leak'],
-    [32, A320Failure.YellowBrakeAccumulatorGasLeak, 'Yellow brake accumulator gas leak'],
-
-    [34, A320Failure.RadioAltimeter1, 'RA 1'],
-    [34, A320Failure.RadioAltimeter2, 'RA 2'],
-]);
+const createOrchestrator = (failures: FailureDefinition[]) => new FailuresOrchestrator('A32NX', failures);
 
 const Context = React.createContext<FailuresOrchestratorContext>({
     allFailures: [],
@@ -122,8 +24,12 @@ const Context = React.createContext<FailuresOrchestratorContext>({
     deactivate: () => Promise.resolve(),
 });
 
-export const FailuresOrchestratorProvider = ({ children }) => {
-    const [orchestrator] = useState(createOrchestrator);
+export interface FailuresOrchestratorProviderProps {
+    failures: FailureDefinition[],
+}
+
+export const FailuresOrchestratorProvider: React.FC<PropsWithChildren<FailuresOrchestratorProviderProps>> = ({ failures, children }) => {
+    const [orchestrator] = useState(() => createOrchestrator(failures));
 
     const [allFailures] = useState(() => orchestrator.getAllFailures());
     const [activeFailures, setActiveFailures] = useState<Set<number>>(() => new Set<number>());
