+++ fbw-common/src/systems/shared/src/failures/communication/callback-reader.ts
