+++ fbw-common/src/systems/shared/src/failures/communication/reader.ts
