+++ fbw-common/src/systems/shared/src/failures/communication/readers/queued-sim-var-reader.ts
