+++ fbw-common/src/systems/shared/src/failures/communication/readers/transaction-reader.spec.ts
