+++ fbw-common/src/systems/shared/src/failures/communication/sim-var-reader-writer.spec.ts
