+++ fbw-common/src/systems/shared/src/failures/communication/sim-var-reader-writer.ts
