+++ fbw-common/src/systems/shared/src/failures/communication/writer.ts
