+++ fbw-common/src/systems/shared/src/failures/communication/writers/index.ts
