+++ fbw-common/src/systems/shared/src/failures/communication/writers/queued-sim-var-writer.ts
