+++ fbw-common/src/systems/shared/src/failures/communication/writers/transaction-writer.ts
