+++ fbw-common/src/systems/shared/src/failures/index.ts
@@ -1,0 +1,3 @@
+export { FailuresConsumer } from './failures-consumer';
+export { FailuresOrchestrator, FailureDefinition } from './failures-orchestrator';
+export type { Failure } from './failures-orchestrator';
