+++ fbw-common/src/systems/shared/src/failures/sim-vars.ts
