+++ fbw-common/src/systems/shared/src/index.ts
@@ -1,3 +1,5 @@
+export * from './failures';
+export * from './navigraph';
 export * from './ApproachUtils';
 export * from './Arinc429ConsumerSubject';
 export * from './Arinc429RegisterSubject';
@@ -15,6 +17,7 @@ export * from './aircraftTypeCheck';
 export * from './arinc429';
 export * from './array';
 export * from './ata';
+export * from './AutoCallOuts';
 export * from './bitFlags';
 export * from './config';
 export * from './logic';
